requirements.txt
aiohttp==3.5.4
APScheduler==3.6.3
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
discord.py==1.2.5
idna==2.8
multidict==4.6.1
pickleDB==0.9.2
pkg-resources==0.0.0
python-dotenv==0.10.3
pytz==2019.3
redis==3.3.11
six==1.13.0
tzlocal==2.0.0
websockets==6.0
yarl==1.3.0
